refactor: http 관련 메소드/model webserver 하위로 이동
