+++ src/main/java/service/FileService.java
@@ -1,12 +1,8 @@
 package service;
 
-import webserver.model.Request;
-
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
-import static http.HttpUtil.*;
-import static service.SessionService.getUserIdBySid;
 
 public class FileService {
