+++ src/main/java/webserver/RequestHandler.java
@@ -3,9 +3,9 @@ package webserver;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import router.Router;
-import webserver.model.Request;
-import webserver.model.Request.Method;
-import webserver.model.Response;
+import webserver.http.model.Request;
+import webserver.http.model.Request.Method;
+import webserver.http.model.Response;
 
 import java.io.*;
 import java.net.Socket;
@@ -14,8 +14,8 @@ import java.nio.charset.StandardCharsets;
 import java.util.HashMap;
 import java.util.Map;
 
-import static http.HttpUtil.*;
-import static http.HttpParser.*;
+import static webserver.http.HttpUtil.*;
+import static webserver.http.HttpParser.*;
 import static service.SessionService.isSessionValid;
 
 public class RequestHandler implements Runnable {
