+++ src/main/java/webserver/http/HttpUtil.java
@@ -1,4 +1,4 @@
-package http;
+package webserver.http;
 
 import java.io.BufferedReader;
 import java.io.IOException;
